Convert to .netcore
This moves the target to .Net Standard 1.6.

We probably only need 1.4 for this to work, but I'm teething
with tooling issues on .Net Core 2.0 preview.

The old test suite remains and the new one is simply an XUnit
port of the MS unit test framework backed Eleven41.Skip32.UnitTests
project.